avoid two compilation warnings
git-svn-id: svn+ssh://svn.in2p3.fr/class@768 0e7d625b-0364-4367-a6be-d5be4a48d228

+++ source/trunk/Model/Equivalence/EQM_FBR_MLP_Keff_BOUND.cxx
@@ -75,7 +75,7 @@ EQM_FBR_MLP_Keff_BOUND::EQM_FBR_MLP_Keff_BOUND(string TMVAWeightPath,  int NumOf
 	INFO << "\tThe TMVA (weight | information) files are :" << endl;
 	INFO << "\t" << "( " << fTMVAWeightPath << " | " << fInformationFile << " )" << endl;
 	INFO << "Time (s) :"<<endl;
-	for(int i=0;i<fMLP_Time.size();i++)
+	for(int i=0;i< (int) fMLP_Time.size();i++)
 			INFO<<fMLP_Time[i]<<endl;
 	INFO<<endl;
 	INFO<<"Z A I Name (input MLP) :"<<endl;
@@ -135,7 +135,7 @@ EQM_FBR_MLP_Keff_BOUND::EQM_FBR_MLP_Keff_BOUND(CLASSLogger* log, string TMVAWeig
 	INFO << "\tThe TMVA (weight | information) files are :" << endl;
 	INFO << "\t" << "( " << fTMVAWeightPath << " | " << fInformationFile << " )" << endl;
 	INFO << "Time (s) :"<<endl;
-	for(int i=0;i<fMLP_Time.size();i++)
+	for(int i=0;i< (int) fMLP_Time.size();i++)
 			INFO<<fMLP_Time[i]<<endl;
 	INFO<<endl;
 	INFO<<"Z A I Name (input MLP) :"<<endl;
